Wallets

+++ database/migrations/2017_07_31_022725_create_transfers_money_table.php
@@ -16,7 +16,7 @@ class CreateTransfersMoneyTable extends Migration
             $table->increments('id');
             $table->integer('transfer_wallet');
             $table->integer('receive_wallet');
-            $table->integer('amount');
+            $table->bigInteger('amount');
             $table->integer('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->timestamps();
